+++ src/index.jsx
@@ -1,3 +1,3 @@
 // We need to include our components so they are included in the bundle
 import './components/contributions-table';
-import './components/search-query';
+import './components/SearchQuery/search-query';
